# api/User-openapi.yaml
openapi: 3.1.0
info:
  title: User Service Papich API
  description: "Документация API для микросервиса юзеров Папич. Управление юзерами\
    , ролями и сессиями"
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Generated server url
...
paths:
  /users:
    post:
      tags:
        - Юзеры
      summary: Создать (зарегистрировать) юзера
      description: Создает нового юзера
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterDto"
        required: true
      responses:
        "201":
          description: Успешное создание (регистрация) юзера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
...
  /users/links:
    patch:
      tags:
        - Юзеры
      summary: Добавить юзеру ссылку на соцсеть
      description: "Добавляет юзеру ссылку на ВК ли ТГ (имя). Если в теле не указа\
        н ID юзера, то ссылка добавляется текущему юзеру (отправившему запрос)"
      operationId: addLink
      parameters:
        - name: linkType
          in: query
          description: Имя соцсети
          required: true
          schema:
            type: string
            enum:
              - TG
              - VK
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserAddLinkDto"
        required: true
      responses:
        "204":
          description: Успешное добавление ссылки
        "401":
          description: Запрос не авторизован
        "403":
          description: Нет прав управлять юзером
        "404":
          description: Юзер не найден
        "409":
          description: Кто-то изменил юзера во время выполнения запроса
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
      security:
        - bearerAuth: [ ]
...
  /users/profile:
    get:
      tags:
        - Юзеры
      summary: Получить инфу о текущем юзере
      description: Возвращает всю инфу о текущем юзере (отправившем запрос)
      operationId: getUser
      responses:
        "200":
          description: Успешное получение инфы о юзере
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        "401":
          description: Запрос не авторизован
        "404":
          description: Юзер не найден
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
      security:
        - bearerAuth: [ ]
  /users/get:
    get:
      tags:
        - Юзеры
      summary: Получить список из части ФИО
      description: "Возвращает список из имен или фамилий, или отчеств по параметр\
        ам. Если указано имя, то возвращается список их похожих имен, аналогично с\
        \ фамилией и отчеством. Пример: firstName=ва, ответ=[Вадим, Вован]"
      operationId: getAllUsersWithName
      parameters:
        - name: firstName
          in: query
          description: Имя
          required: false
          schema:
            type: string
        - name: lastName
          in: query
          description: Фамилия
          required: false
          schema:
            type: string
        - name: middleName
          in: query
          description: Отчество
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Успешное получение частей ФИО
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "401":
          description: Запрос не авторизован
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
      security:
        - bearerAuth: [ ]
  /users/get/with/roles/{id}:
    get:
      tags:
        - Юзеры
      summary: Получить инфу с ролями о юзере по ID
      description: Возвращает инфу с ролями о юзере по ID
      operationId: getUserWithRoles
      parameters:
        - name: id
          in: path
          description: ID юзера
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Успешное получение инфы о юзере
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseWithRoleDto"
        "401":
          description: Запрос не авторизован
        "404":
          description: Юзер не найден
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
      security:
        - bearerAuth: [ ]
  /users/get/room/{userId}:
    get:
      tags:
        - Юзеры
      summary: Получить номер комнаты юзера
      description: Возвращает номер комнаты юзера с указанным ID
      operationId: getRoomNumber
      parameters:
        - name: userId
          in: path
          description: ID юзера
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Успешное получение инфы о юзерах
          content:
            application/json:
              schema:
                type: string
        "401":
          description: Запрос не авторизован
        "404":
          description: Юзер не найден
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
      security:
        - bearerAuth: [ ]
  /users/get/by/role:
    get:
      tags:
        - Юзеры
      summary: Получить список из краткой инфы о юзерах с ролью
      description: "Постранично возвращает список из краткой инфы о юзерах с ролью\
        . Если указать параметр onMyFloor, то вернется список для юзеров с того же\
        \ этажа, что и текущий (отправивший запрос). Пагинация при этом игнорирует\
        ся"
      operationId: getUsersByRole
      parameters:
        - name: role
          in: query
          description: Роль
          required: true
          schema:
            type: string
            enum:
              - ADMINISTRATION
              - HOSTEL_SUPERVISOR
              - FLOOR_SUPERVISOR
              - RESPONSIBLE_KITCHEN
              - RESPONSIBLE_HALL
              - RESPONSIBLE_GYM
              - WORKER_GYM
              - RESPONSIBLE_FIRE_SAFETY
              - RESPONSIBLE_SANITARY
              - RESPONSIBLE_INTERNET
              - RESPONSIBLE_SOOP
              - WORKER_FIRE_SAFETY
              - WORKER_SANITARY
              - WORKER_SOOP
              - STUDENT
        - name: page
          in: query
          description: Номер страницы
          required: true
          schema:
            type: integer
            format: int32
        - name: size
          in: query
          description: Количество записей на странице
          required: true
          schema:
            type: integer
            format: int32
        - name: onMyFloor
          in: query
          description: Указывает на поиск в рамках этажа текущего юзера. Отключает пагинацию
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Успешное получение инфы о юзерах
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserShortResponseDto2"
        "401":
          description: Запрос не авторизован
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
      security:
        - bearerAuth: [ ]
  /users/get/by/name:
    get:
      tags:
        - Юзеры
      summary: Получить список из краткой инфы о юзерах с ФИО
      description: Постранично возвращает список из краткой инфы о юзерах с похожим
        ФИО
      operationId: getUsersByName
      parameters:
        - name: name
          in: query
          description: ФИО
          required: false
          schema:
            type: string
            default: ""
        - name: page
          in: query
          description: Номер страницы
          required: true
          schema:
            type: integer
            format: int32
        - name: size
          in: query
          description: Количество записей на странице
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Успешное получение инфы о юзерах
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserShortResponseDto2"
        "401":
          description: Запрос не авторизован
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
      security:
        - bearerAuth: [ ]
  /users/get/by/name/without/role:
    get:
      tags:
        - Юзеры
      summary: Получить список из краткой инфы о юзерах с ФИО и без роли
      description: Постранично возвращает список из краткой инфы о юзерах с похожим
        ФИО и без роли
      operationId: getUsersByNameWithoutRole
      parameters:
        - name: name
          in: query
          description: ФИО
          required: false
          schema:
            type: string
            default: ""
        - name: role
          in: query
          description: Роль
          required: true
          schema:
            type: string
            enum:
              - ADMINISTRATION
              - HOSTEL_SUPERVISOR
              - FLOOR_SUPERVISOR
              - RESPONSIBLE_KITCHEN
              - RESPONSIBLE_HALL
              - RESPONSIBLE_GYM
              - WORKER_GYM
              - RESPONSIBLE_FIRE_SAFETY
              - RESPONSIBLE_SANITARY
              - RESPONSIBLE_INTERNET
              - RESPONSIBLE_SOOP
              - WORKER_FIRE_SAFETY
              - WORKER_SANITARY
              - WORKER_SOOP
              - STUDENT
        - name: page
          in: query
          description: Номер страницы
          required: true
          schema:
            type: integer
            format: int32
        - name: size
          in: query
          description: Количество записей на странице
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Успешное получение инфы о юзерах
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserShortResponseDto2"
        "401":
          description: Запрос не авторизован
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
      security:
        - bearerAuth: [ ]
  /users/get/by/name/with/role:
    get:
      tags:
        - Юзеры
      summary: Получить список из краткой инфы о юзерах с ФИО и ролью
      description: Постранично возвращает список из краткой инфы о юзерах с похожим
        ФИО и ролью
      operationId: getUsersByNameWithRole
      parameters:
        - name: name
          in: query
          description: ФИО
          required: false
          schema:
            type: string
            default: ""
        - name: role
          in: query
          description: Роль
          required: true
          schema:
            type: string
            enum:
              - ADMINISTRATION
              - HOSTEL_SUPERVISOR
              - FLOOR_SUPERVISOR
              - RESPONSIBLE_KITCHEN
              - RESPONSIBLE_HALL
              - RESPONSIBLE_GYM
              - WORKER_GYM
              - RESPONSIBLE_FIRE_SAFETY
              - RESPONSIBLE_SANITARY
              - RESPONSIBLE_INTERNET
              - RESPONSIBLE_SOOP
              - WORKER_FIRE_SAFETY
              - WORKER_SANITARY
              - WORKER_SOOP
              - STUDENT
        - name: page
          in: query
          description: Номер страницы
          required: true
          schema:
            type: integer
            format: int32
        - name: size
          in: query
          description: Количество записей на странице
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: Успешное получение инфы о юзерах
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserShortResponseDto2"
        "401":
          description: Запрос не авторизован
        "500":
          description: Неизвестная ошибка сервера
        "400":
          description: "Неверный запрос от клиента, нарушение ограничений запроса\
            \ (тело, параметры)"
      security:
        - bearerAuth: [ ]
...
components:
  schemas:
    UserRegisterDto:
      type: object
      description: Тело запроса на регистрацию юзера
      properties:
        firstName:
          type: string
          description: Имя
          example: Вадим
          maxLength: 50
          minLength: 1
          pattern: "^[А-ЯЁ][а-яё]{1,50}$"
        lastName:
          type: string
          description: Фамилия
          example: Гудко
          maxLength: 50
          minLength: 1
          pattern: "^[А-ЯЁ][а-яё]{1,50}$"
        middleName:
          type: string
          description: Отчество (опционально)
          example: Макарыч
          maxLength: 50
          minLength: 0
          pattern: "^$|^([А-ЯЁ][а-яё]{0,50})$"
        email:
          type: string
          format: email
          description: Имейл
          maxLength: 100
          minLength: 5
        phoneNumber:
          type: string
          description: Номер телефона (опционально)
          example: 79137235412
        password:
          type: string
          description: Пароль
          minLength: 1
        roomNumber:
          type: string
          description: Номер комнаты
          minLength: 1
          pattern: "^\\d{3}[АБаб]?$"
      required:
        - email
        - firstName
        - lastName
        - password
        - roomNumber
    UserResponseDto:
      type: object
      description: "Ответ со всей инфой о юзере, кроме пароля"
      properties:
        id:
          type: string
          format: uuid
          description: ID
        firstName:
          type: string
          description: Имя
        lastName:
          type: string
          description: Фамилия
        middleName:
          type: string
          description: Отчество
        email:
          type: string
          description: Имейл
        phoneNumber:
          type: string
          description: Номер телефона
        roomNumber:
          type: string
          description: Номер комнаты
        tgLink:
          type: string
          description: Имя в ТГ
        vkLink:
          type: string
          description: Имя в ВК
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT